+++ projects/jsf/src/lib/validator.service.spec.ts
@@ -2,7 +2,7 @@ import { FormControl, ValidatorFn, Validators } from '@angular/forms';
 import { FormDataItem, FormDataItemType } from './models/form-data-item';
 import { IntegerDataItem, IntegerRangeOptions } from './models/integer-data-item';
 import { StringDataItem, StringFormat, StringLengthOptions } from './models/string-data-item';
-import { ValidatorService } from './validator.service';
+import { URI_REGEX, ValidatorService } from './validator.service';
 import Spy = jasmine.Spy;
 
 describe('ValidatorService', () => {
@@ -236,6 +236,31 @@ describe('ValidatorService', () => {
           it('validates for a single uri', () => {
             control.setValue('https://test.Okta.com/api');
             expect(validatorFn(control)).toBeNull();
+          });
+
+          it('it correctly validates uris against the uri format regex', () => {
+            const urls: [string, boolean][] = [
+              ['test.com', true],
+              ['https://test.com', true],
+              ['https://test.com:4200', true],
+              ['ftp://test.com', true],
+              ['test.com?query=testquery', true],
+              ['test-hyphen.com', true],
+              ['http://test.com', true],
+              ['http://test.co', true],
+              ['test.com/🤔', false],
+              ['test.com/text\u0002.com', false],
+              ['http://test.c', false],
+              ['test', false],
+              ['http://test .com', false],
+              ['https//test.com', false]
+            ];
+
+            for (const testSet of urls) {
+              const url = testSet[0];
+              const expectedValidity = testSet[1];
+              expect(URI_REGEX.test(url)).toBe(expectedValidity, `url: ${url} expectedValidity: ${expectedValidity}`);
+            }
           });
         });